Create Ch13-GroupsTeamsReport.PS1
A script to check the activity of Office 365 Groups and Teams and report the groups and teams that might be deleted because they're not used. We check the group mailbox to see what the last time a conversation item was added to the Inbox folder. 
Another check sees whether a low number of items exist in the mailbox, which would show that it's not being used. We also check the group document library in SharePoint Online to see whether it exists or has been used in the last 90 days. And we check Teams compliance items to figure out if any chatting is happening.